refactor: split function groups into separate files
